Windows:  More name normalization shenanigans
This is quite, quite, hideous.  Under Windows we have to prefix "DOS device" type paths with "/" so that the 'prefix with "file://"' paradigm works.  So IDP_HOME and its friends have to be 'normalized'.  However java.nio.Paths doesn't understand "\c:", as a path, it only understands "c:" (don't even ask).  So for Strings that get passed into java.nio.Paths we need the non-normalized name.  We achieve this by setting the property as normalized, but returning the non normalized string.